Clean up and consolidate article API directories (#54509)
