+++ src/article-api/README.md
@@ -1,0 +1,10 @@
+# Article API
+
+This subject folder contains the code for the Article API endpoints:
+- `/api/pagelist`
+- `/api/article/body`
+- `/api/article/meta`
+
+## What it does
+
+Article API endpoints allow consumers to query GitHub Docs for listings of current articles, and for specific article information.
