+++ src/frame/middleware/api.ts
@@ -6,7 +6,7 @@ import anchorRedirect from '@/rest/api/anchor-redirect.js'
 import aiSearch from '@/search/middleware/ai-search'
 import search from '@/search/middleware/search-routes.js'
 import pageInfo from '@/pageinfo/middleware'
-import pageList from '@/pagelist/middleware'
+import pageList from '@/article-api/middleware/pagelist'
 import webhooks from '@/webhooks/middleware/webhooks.js'
 import { ExtendedRequest } from '@/types'
 import { noCacheControl } from './cache-control'
