--- src/pagelist/README.md
@@ -1,7 +1,0 @@
-# Pagelist
-
-This subject folder contains the code for the `/api/pagelist` endpoint.
-
-## What it does
-
-The `/api/pagelist` endpoint provides a flat structure of all the pages on GitHub Docs.
